+++ .travis_key.sh
@@ -2,15 +2,18 @@
 
 set -e
 
-if [ -z ${TRAVIS} ]; then
+if [ ! "${TRAVIS:-}" = 'true' ]; then
+  echo 'This is not running on Travis CI. Exiting!'
   exit 0
 fi
 
-if [ "${TRAVIS_PULL_REQUEST}" = "true" ]; then
+if [ ! "${TRAVIS_PULL_REQUEST:-}" = 'false' ]; then
+  echo 'Skipping deploy step for pull request.'
   exit 0
 fi
 
-if [ "${SKIP_DEPLOY}" = "true" ] && [ "${REQUIRE_KEY}" != "true" ]; then
+if [ "${SKIP_DEPLOY:-}" = "true" ] && [ "${REQUIRE_KEY:-}" != "true" ]; then
+  echo 'Skip deploy explicitly set.'
   exit 0
 fi